requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.11.1
python-dotenv==1.0.0
pydantic==2.4.1
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1
python-dateutil==2.9.0.post0
pandas==2.2.2
PyYAML==6.0.1
httpx==0.25.1
requests==2.31.0
twilio==8.10.0
beautifulsoup4==4.12.2
lxml==4.9.3
geopy==2.3.0
openpyxl==3.1.2